Use transactions to save/retrieve posts
This was initially an attempt to fix performance problems, because I was
reading about how SQLite can only do one write transaction, but the
problem turned out to be something else and was fixed by using an index
with e879127.

But transactions are still good, and maybe it also helps to tell the
database which queries are read-only, and which will write.