Count views
Not sure if we really want this, but another measure of how many
requests are happening.  The important measure so far is that no
personal data is tracked, in particular which pages are viewed.

It is strange how tempting it is to want to track more data.  I will
still not do it.  And stats are still (and will continue to be) off by
default.