Only return from cache if feedError is non-empty
The db would return NULL for entries where the `error` field was not set
(yet) and thus all entries that had the field set but empty (i.e. every
post) would be cached.  This should fix that and only cache the feeds
that actually have errors, and leave the others alone.