Fix tiktok caching
Lookups happen by author although the primary key is source + _name_ +
id.  The indices are on the author as well, so I am not even sure
anymore why there is a name field.  Maybe name was the programmatic id
at some point and author was supposed to be the pretty name and at some
point I just mixed them up?