Improve last post/next page handling
This very likely not the "final" fix to this, it's just an improvement.

`lastPost` is now only calculated based on posts actually displayed
and the default post limit now matches the one used to select from the
db, which should improve the logic a bit.

This is another case... where tests would help!  🙄